fix: RAT reliability (C5ROL, retry/diag infrastructure), auto-start tests, CI coverage, un-nest routes, split RAT into rat/ package
